// packages/client/src/types.d.ts
type RequestInteractionTypes = {
  read: "read-request";
  vread: "vread-request";
  update: "update-request";
  patch: "patch-request";
  delete: "delete-request";
  history: "history-request";
  create: "create-request";
  search: "search-request";
  capabilities: "capabilities-request";
  batch: "batch-request";
  transaction: "transaction-request";
  invoke: "invoke-request";
};
// ...
export type InstanceInteraction = {
  level: RequestLevel["instance"];
  resourceType: string;
  id: id;
};

export type TypeInteraction = {
  level: RequestLevel["type"];
  resourceType: string;
};

export type SystemInteraction = {
  level: RequestLevel["system"];
};

export type ReadRequest = InstanceInteraction & {
  type: RequestInteractionTypes["read"];
};

export type VersionReadRequest = InstanceInteraction & {
  type: RequestInteractionTypes["vread"];
  versionId: string;
};

export type UpdateRequest = InstanceInteraction & {
  type: RequestInteractionTypes["update"];
  body: Resource;
};

// TODO - implement patch type
export type PatchRequest = InstanceInteraction & {
  type: RequestInteractionTypes["patch"];
  body: Object;
};

export type DeleteRequest = InstanceInteraction & {
  type: RequestInteractionTypes["delete"];
};

export type HistoryInstanceRequest = InstanceInteraction & {
  type: RequestInteractionTypes["history"];
};

export type CreateRequest = TypeInteraction & {
  type: RequestInteractionTypes["create"];
  body: Resource;
};

export type TypeSearchRequest = TypeInteraction & {
  parameters: ParsedParameter<string | number>[];
  type: RequestInteractionTypes["search"];
};

export type TypeHistoryRequest = TypeInteraction & {
  type: RequestInteractionTypes["history"];
};

export type CapabilitiesRequest = SystemInteraction & {
  type: RequestInteractionTypes["capabilities"];
};

export type BatchRequest = SystemInteraction & {
  type: RequestInteractionTypes["batch"];
  body: Bundle;
};

export type TransactionRequest = SystemInteraction & {
  type: RequestInteractionTypes["transaction"];
  body: Bundle;
};

export type SystemHistoryRequest = SystemInteraction & {
  type: RequestInteractionTypes["history"];
};

export type SystemSearchRequest = SystemInteraction & {
  parameters: ParsedParameter<string | number>[];
  type: RequestInteractionTypes["search"];
};

export type InvokeInstanceRequest = InstanceInteraction & {
  type: RequestInteractionTypes["invoke"];
  operation: string;
  body: Parameters;
};

export type InvokeTypeRequest = TypeInteraction & {
  type: RequestInteractionTypes["invoke"];
  operation: string;
  body: Parameters;
};

export type InvokeSystemRequest = SystemInteraction & {
  type: RequestInteractionTypes["invoke"];
  operation: string;
  body: Parameters;
};

export type FHIRRequest =
  | InvokeInstanceRequest
  | InvokeTypeRequest
  | InvokeSystemRequest
  | ReadRequest
  | VersionReadRequest
  | UpdateRequest
  | PatchRequest
  | DeleteRequest
  | HistoryInstanceRequest
  | CreateRequest
  | TypeSearchRequest
  | TypeHistoryRequest
  | CapabilitiesRequest
  | BatchRequest
  | TransactionRequest
  | SystemHistoryRequest
  | SystemSearchRequest;